+++ CHANGELOG.md
@@ -4,6 +4,13 @@ Updates should follow the [Keep a CHANGELOG](https://keepachangelog.com/) princi
 
 ## [Unreleased][unreleased]
 
+## [0.3.0] - 2022-05-17
+
+### Changed
+
+- Added support for namespaced variable names.
+- Removed support for `SHIELD*` vars. Use `PUBLIC` instead.
+
 ## [0.2.1] - 2022-03-30
 
 ### Fixed
@@ -141,7 +148,8 @@ Updates should follow the [Keep a CHANGELOG](https://keepachangelog.com/) princi
 
 **Initial release!**
 
-[unreleased]: https://github.com/unleashedtech/dotenv-drupal/compare/0.2.1...main
+[unreleased]: https://github.com/unleashedtech/dotenv-drupal/compare/0.3.0...main
+[0.3.0]: https://github.com/unleashedtech/dotenv-drupal/compare/0.2.1...0.3.0
 [0.2.1]: https://github.com/unleashedtech/dotenv-drupal/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/unleashedtech/dotenv-drupal/compare/0.1.17...0.2.0
 [0.1.17]: https://github.com/unleashedtech/dotenv-drupal/compare/0.1.16...0.1.17